dagster-project/pyproject.toml
[project]
name = "dagster_project"
requires-python = ">=3.9,<3.14"
version = "0.1.0"
dependencies = [
    "dagster==1.11.15",
    "dagster-dbt==0.27.15",
    "dagster-webserver",
    "dbt-bigquery==1.10.2",
    "dagster-gcp==0.27.15",
    "google-cloud-bigquery==3.38.0"
]

[project.optional-dependencies]
dev = [
    "dagster-dg-cli"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.dg]
directory_type = "project"

[tool.dg.project]
root_module = "dagster_project"
registry_modules = [
    "dagster_project.components.*",
]

[tool.dagster]
module_name = "dagster_project.definitions"